Rename endpoint name
