+++ FitBridge_Application/Features/Transactions/GetCurrentUserTransactions/GetCurrentUserTransactionsHandler.cs
@@ -3,26 +3,27 @@ using FitBridge_Application.Dtos;
 using FitBridge_Application.Dtos.Transactions;
 using FitBridge_Application.Interfaces.Repositories;
 using FitBridge_Application.Interfaces.Utils;
-using FitBridge_Application.Specifications.Transactions;
-using FitBridge_Application.Specifications.Transactions.GetTransactionByPtId;
+using FitBridge_Application.Specifications.Transactions.GetCurrentUserTransactions;
+using FitBridge_Domain.Entities.Identity;
 using FitBridge_Domain.Entities.Orders;
 using FitBridge_Domain.Exceptions;
 using MediatR;
 using Microsoft.AspNetCore.Http;
 
-namespace FitBridge_Application.Features.Transactions.GetFreelancePtTransactions
+namespace FitBridge_Application.Features.Transactions.GetCurrentUserTransactions
 {
-    internal class GetFreelancePtTransactionsQueryHandler(
+    internal class GetCurrentUserTransactionsHandler(
         IUserUtil userUtil,
         IHttpContextAccessor httpContextAccessor,
         IUnitOfWork unitOfWork,
-        IMapper mapper) : IRequestHandler<GetFreelancePtTransactionsQuery, PagingResultDto<GetTransactionsDto>>
+        IMapper mapper) : IRequestHandler<GetCurrentUserTransactionsQuery, PagingResultDto<GetTransactionsDto>>
     {
-        public async Task<PagingResultDto<GetTransactionsDto>> Handle(GetFreelancePtTransactionsQuery request, CancellationToken cancellationToken)
+        public async Task<PagingResultDto<GetTransactionsDto>> Handle(GetCurrentUserTransactionsQuery request, CancellationToken cancellationToken)
         {
             var accountId = userUtil.GetAccountId(httpContextAccessor.HttpContext)
-                    ?? throw new NotFoundException(nameof(Transaction));
-            var spec = new GetTransactionByPtIdSpec(request.Parameters, accountId, includeWithdrawRequest: true);
+                    ?? throw new NotFoundException(nameof(ApplicationUser));
+
+            var spec = new GetCurrentUserTransactionsSpec(request.Parameters, accountId, includeWithdrawRequest: true);
             var transactions = await unitOfWork.Repository<Transaction>()
                 .GetAllWithSpecificationProjectedAsync<GetTransactionsDto>(spec, mapper.ConfigurationProvider);
             var totalCount = await unitOfWork.Repository<Transaction>()
