+++ FitBridge_Application/Specifications/Transactions/GetCurrentUserTransactions/GetCurrentUserTransactionsParam.cs
@@ -1,0 +1,6 @@
+﻿namespace FitBridge_Application.Specifications.Transactions.GetCurrentUserTransactions
+{
+    public class GetCurrentUserTransactionsParam : BaseParams
+    {
+    }
+}
