--- FitBridge_Application/Specifications/Transactions/GetTransactionByPtId/GetTransactionByPtIdParam.cs
@@ -1,6 +1,0 @@
-﻿namespace FitBridge_Application.Specifications.Transactions.GetTransactionByPtId
-{
-    public class GetTransactionByPtIdParam : BaseParams
-    {
-    }
-}
